+++ samples/Messaging/MiniChatEx/MiniChatEx.cpp
@@ -52,7 +52,7 @@ PostOneMessage (
 HRESULT
 ReceiveOneMessage (
     _In_ LMQ_CHANNEL Channel,
-    _In_ int ExpectedPayloadSizeBytes,
+    _In_ SIZE_T ExpectedPayloadSizeBytes,
     _In_ BOOL PrintData
     ) noexcept;
 
@@ -202,7 +202,7 @@ PostOneMessage (
 HRESULT
 ReceiveOneMessage(
     _In_ LMQ_CHANNEL Channel,
-    _In_ int ExpectedPayloadSizeBytes,
+    _In_ SIZE_T ExpectedPayloadSizeBytes,
     _In_ BOOL PrintData
 ) noexcept
 {
